Giant PR that adds many features, including
* Support for ESP32-S3 including USB CDC serial ports
* WebDAV support for managing FluidNC files from host operating systems, e.g. Windows Explorer
* Asynchronous TCP and Web Server for better performance and stability
* Parallel Delta kinematics

+++ FluidNC/src/FileStream.h
@@ -47,8 +47,8 @@ public:
 
     size_t readBytes(char* buffer, size_t length) { return read((uint8_t*)buffer, length); }
 
-    size_t read(char* buffer, size_t length);  // read chars from stream into buffer
-    size_t read(uint8_t* buffer, size_t length) { return read((char*)buffer, length); }
+    int read(char* buffer, size_t length);  // read chars from stream into buffer
+    int read(uint8_t* buffer, size_t length) { return read((char*)buffer, length); }
 
     size_t write(uint8_t c) override;
     size_t write(const uint8_t* buffer, size_t length) override;
