+++ FluidNC/src/UTF8.h
@@ -8,12 +8,12 @@
 
 class UTF8 {
 private:
-    uint32_t _num = 0;
-    int      _state = 0;
+    uint32_t _num   = 0;
+    uint8_t  _state = 0;
 
 public:
     // Byte-at-a-time decoder.  Returns -1 for error, 1 for okay, 0 for keep trying
-    int decode(uint8_t ch, uint32_t& value);
+    int8_t decode(uint8_t ch, uint32_t& value);
 
     // Vector-of-bytes decoder.  Returns true if the vector contains
     // a well-formed UTF8 sequence.
